Ignore HTML entities parse error when loading xml
We use libxml to load the xhtml and replace each resource path with a
custom epub:// path. Libxml shows error in the stderr when the xhtml has
html entities:

Entity: line 21: parser error : Entity 'copy' not defined

We don't mind html entities in this case, because we only want to replace
resources path so I've replaced xmlRecoverMemory by xmlReadMemory to pass
the parser options XML_PARSE_NOWARNING and XML_PARSE_NOERROR and we won't
see the error in the console anymore.