tests: Fix crash on exit
When the main window is destroyed, and the view widget is as well, we'd
receive a property change ("notify") signal for the parent widget of the
view changing. Except that we're in the middle of destroying that view.

Disconnect from the "notify" signal before exiting, so that we don't
crash on exit.