Helper functions to convert from res to chapter
The `gepub_doc_resource_uri_to_chapter` and
`gepub_doc_resource_id_to_chapter` methods are useful to convert a
resource ID or URI to a chapter number and then use in the
`gepub_doc_set_chapter` method, to navigate.

This new helper functions can be combined with the new
`gepub_doc_get_toc` to navigate in the document.

See #6